# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Echo Template API
  description: Echo Template 项目的 RESTful API 文档
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: 开发环境
  - url: https://api.example.com
    description: 生产环境
security:
  - BearerAuth: []
...
paths:
...
  /api/v1/users/{id}:
    get:
      tags:
        - 用户
      summary: 获取用户信息
      description: 根据用户ID获取用户详细信息
      operationId: getUserById
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      responses:
        '200':
          description: 获取用户信息请求响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'

    put:
      tags:
        - 用户
      summary: 更新用户信息
      description: 根据用户ID更新用户信息
      operationId: updateUserById
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserInput'
      responses:
        '200':
          description: 更新用户信息请求响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'

    delete:
      tags:
        - 用户
      summary: 删除用户
      description: 根据用户ID删除用户
      operationId: deleteUserById
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/UserIdParam'
      responses:
        '200':
          description: 删除用户请求响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
...
  /api/v1/roles:
    get:
      tags:
        - 角色管理
      summary: 获取角色列表
      description: 分页获取角色列表，支持状态过滤和搜索
      operationId: getRoles
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: status
          in: query
          description: 角色状态过滤
          required: false
          schema:
            type: string
            enum: [active, inactive]
        - name: search
          in: query
          description: 搜索关键词（角色名称、代码或描述）
          required: false
          schema:
            type: string
      responses:
        '200':
          description: 获取角色列表请求响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RolesResponse'

    post:
      tags:
        - 角色管理
      summary: 创建角色
      description: 创建新的角色
      operationId: createRole
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoleInput'
      responses:
        '200':
          description: 创建角色请求响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'
  /api/v1/roles/{id}:
    get:
      tags:
        - 角色管理
      summary: 获取角色详情
      description: 根据角色ID获取角色详细信息
      operationId: getRoleById
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          description: 角色ID
          required: true
          schema:
            type: string
            minLength: 26
            maxLength: 26
      responses:
        '200':
          description: 获取角色详情请求响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'

    put:
      tags:
        - 角色管理
      summary: 更新角色
      description: 更新角色信息
      operationId: updateRole
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          description: 角色ID
          required: true
          schema:
            type: string
            minLength: 26
            maxLength: 26
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRoleInput'
      responses:
        '200':
          description: 更新角色请求响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleResponse'

    delete:
      tags:
        - 角色管理
      summary: 删除角色
      description: 删除指定角色（软删除）
      operationId: deleteRole
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          description: 角色ID
          required: true
          schema:
            type: string
            minLength: 26
            maxLength: 26
      responses:
        '200':
          description: 删除角色请求响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /api/v1/roles/assign:
    post:
      tags:
        - 角色管理
      summary: 分配角色给用户
      description: 为用户分配一个或多个角色
      operationId: assignRoles
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignRoleInput'
      responses:
        '200':
          description: 分配角色请求响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /api/v1/roles/revoke:
    post:
      tags:
        - 角色管理
      summary: 撤销用户角色
      description: 撤销用户的一个或多个角色
      operationId: revokeRoles
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RevokeRoleInput'
      responses:
        '200':
          description: 撤销角色请求响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /api/v1/roles/users/{user_id}:
    get:
      tags:
        - 角色管理
      summary: 获取用户角色列表
      description: 获取指定用户的所有角色
      operationId: getUserRoles
      security:
        - BearerAuth: []
      parameters:
        - name: user_id
          in: path
          description: 用户ID
          required: true
          schema:
            type: string
            minLength: 26
            maxLength: 26
      responses:
        '200':
          description: 获取用户角色请求响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserRolesResponse'
  /api/v1/roles/{id}/permissions:
    get:
      tags:
        - 角色管理
      summary: 获取角色权限列表
      description: 获取指定角色的所有权限
      operationId: getRolePermissions
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          description: 角色ID
          required: true
          schema:
            type: string
            minLength: 26
            maxLength: 26
      responses:
        '200':
          description: 获取角色权限请求响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RolePermissionsResponse'

    post:
      tags:
        - 角色管理
      summary: 为角色分配权限
      description: 为指定角色分配一个或多个权限
      operationId: assignPermissionsToRole
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          description: 角色ID
          required: true
          schema:
            type: string
            minLength: 26
            maxLength: 26
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignPermissionInput'
      responses:
        '200':
          description: 分配权限请求响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /api/v1/roles/{id}/menus:
    get:
      tags:
        - 角色管理
      summary: 获取角色菜单列表
      description: 获取指定角色的所有菜单
      operationId: getRoleMenus
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          description: 角色ID
          required: true
          schema:
            type: string
            minLength: 26
            maxLength: 26
      responses:
        '200':
          description: 获取角色菜单请求响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleMenuResponse'

    post:
      tags:
        - 角色管理
      summary: 为角色分配菜单
      description: 为指定角色分配一个或多个菜单
      operationId: assignMenusToRole
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          description: 角色ID
          required: true
          schema:
            type: string
            minLength: 26
            maxLength: 26
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignRoleMenusInput'
      responses:
        '200':
          description: 分配菜单请求响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /api/v1/roles/{id}/menus/permissions:
    get:
      tags:
        - 角色管理
      summary: 获取角色菜单权限详情
      description: 获取指定角色的菜单列表及相关权限信息
      operationId: getRoleMenuPermissions
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          description: 角色ID
          required: true
          schema:
            type: string
            minLength: 26
            maxLength: 26
      responses:
        '200':
          description: 获取角色菜单权限请求响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleMenuPermissionResponse'
  /api/v1/roles/users/menus:
    get:
      tags:
        - 角色管理
      summary: 获取用户可访问菜单
      description: 根据用户角色获取用户可访问的菜单列表
      operationId: getUserMenus
      security:
        - BearerAuth: []
      parameters:
        - name: user_id
          in: query
          description: 用户ID
          required: true
          schema:
            type: string
            minLength: 26
            maxLength: 26
        - name: tree_mode
          in: query
          description: 是否返回树形结构
          required: false
          schema:
            type: boolean
            default: true
        - name: only_menu
          in: query
          description: 是否只返回菜单类型（排除按钮）
          required: false
          schema:
            type: boolean
            default: true
      responses:
        '200':
          description: 获取用户菜单请求响应
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserMenuResponse'

  # 权限管理相关路径
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT 访问令牌
  schemas:
...
    ErrorDetail:
      title: 错误详情
      type: object
      description: 错误详情
      properties:
        message:
          type: string
          description: 错误消息
          example: "用户名长度不能小于3"
        location:
          type: string
          description: 错误位置
          example: "username"
        value:
          type: string
          description: 错误的值
          example: "ab"
      required:
        - message

    # 认证相关Schema定义
    
    # 用户注册输入
...
    UpdateUserInput:
      title: 更新用户请求
      type: object
      description: 更新用户输入
      properties:
        username:
          type: string
          description: 用户名
          minLength: 3
          maxLength: 50
          example: "johndoe"
          nullable: true
        email:
          type: string
          format: email
          description: 邮箱地址
          example: "john@example.com"
          nullable: true
        status:
          type: string
          enum: [active, inactive, suspended]
          description: 用户状态
          example: "active"
          nullable: true

    # 修改密码输入
...
    UserResponse:
      title: 用户信息响应
      type: object
      description: 单个用户信息响应
      properties:
        code:
          type: integer
          description: API状态码（0=成功，非0=失败）
          example: 0
        message:
          type: string
          description: 响应消息
          example: "获取用户信息成功"
        data:
          type: object
          description: 用户基本信息
          properties:
            id:
              type: string
              description: 用户ID
              example: "usr_123456"
            username:
              type: string
              description: 用户名
              example: "johndoe"
            email:
              type: string
              format: email
              description: 邮箱地址
              example: "john@example.com"
            status:
              type: string
              enum: [active, inactive, suspended]
              description: 用户状态
              example: "active"
            last_login_at:
              type: string
              format: date-time
              nullable: true
              description: 最后登录时间
              example: "2023-12-31T23:59:59Z"
            created_at:
              type: string
              format: date-time
              description: 创建时间
              example: "2023-01-01T00:00:00Z"
          required:
            - id
            - username
            - email
            - status
            - created_at
        errors:
          type: array
          description: 错误详情列表
          items:
            $ref: '#/components/schemas/ErrorDetail'
        timestamp:
          type: integer
          format: int64
          description: 时间戳
          example: 1640995200
        request_id:
          type: string
          description: 请求ID
          example: "req-123456"
      required:
        - code
        - message
        - data
        - timestamp

    # 用户列表响应
...
    MessageResponse:
      title: 操作结果响应
      type: object
      description: 操作消息响应
      properties:
        code:
          type: integer
          description: API状态码（0=成功，非0=失败）
          example: 0
        message:
          type: string
          description: 响应消息
          example: "操作成功"
        data:
          type: object
          description: 消息数据
          properties:
            message:
              type: string
              description: 消息内容
              example: "操作成功"
          required:
            - message
        errors:
          type: array
          description: 错误详情列表
          items:
            $ref: '#/components/schemas/ErrorDetail'
        timestamp:
          type: integer
          format: int64
          description: 时间戳
          example: 1640995200
        request_id:
          type: string
          description: 请求ID
          example: "req-123456"
      required:
        - code
        - message
        - data
        - timestamp

    # 菜单相关Schema定义

    # 创建菜单输入
...
    CreateRoleInput:
      title: 创建角色请求
      type: object
      description: 创建角色输入
      properties:
        name:
          type: string
          description: 角色名称
          minLength: 1
          maxLength: 50
          example: "管理员"
        code:
          type: string
          description: 角色代码
          pattern: '^[A-Z_]+$'
          minLength: 1
          maxLength: 50
          example: "ADMIN"
        description:
          type: string
          description: 角色描述
          maxLength: 500
          example: "系统管理员角色"
        status:
          type: string
          description: 角色状态
          enum: [active, inactive]
          default: active
          example: "active"
        sort_order:
          type: integer
          description: 排序顺序
          minimum: 0
          example: 1
      required:
        - name
        - code

    # 更新角色输入
    UpdateRoleInput:
      title: 更新角色请求
      type: object
      description: 更新角色输入
      properties:
        name:
          type: string
          description: 角色名称
          minLength: 1
          maxLength: 50
          example: "管理员"
        description:
          type: string
          description: 角色描述
          maxLength: 500
          example: "系统管理员角色"
        status:
          type: string
          description: 角色状态
          enum: [active, inactive]
          example: "active"
        sort_order:
          type: integer
          description: 排序顺序
          minimum: 0
          example: 1

    # 分配角色输入
    AssignRoleInput:
      title: 分配角色请求
      type: object
      description: 分配角色输入
      properties:
        user_id:
          type: string
          description: 用户ID
          minLength: 26
          maxLength: 26
          example: "01HQZX123456789ABCDEFGHIJ"
        role_ids:
          type: array
          description: 角色ID列表
          items:
            type: string
            minLength: 26
            maxLength: 26
          example: ["01HQZX123456789ABCDEFGHIJ", "01HQZX123456789ABCDEFGHIK"]
        expires_at:
          type: string
          format: date-time
          description: 过期时间（可选）
          example: "2024-12-31T23:59:59Z"
        remark:
          type: string
          description: 备注
          maxLength: 500
          example: "临时分配管理员权限"
      required:
        - user_id
        - role_ids

    # 撤销角色输入
    RevokeRoleInput:
      title: 撤销角色请求
      type: object
      description: 撤销角色输入
      properties:
        user_id:
          type: string
          description: 用户ID
          minLength: 26
          maxLength: 26
          example: "01HQZX123456789ABCDEFGHIJ"
        role_ids:
          type: array
          description: 角色ID列表
          items:
            type: string
            minLength: 26
            maxLength: 26
          example: ["01HQZX123456789ABCDEFGHIJ", "01HQZX123456789ABCDEFGHIK"]
      required:
        - user_id
        - role_ids

    # 分配权限输入
    AssignPermissionInput:
      title: 分配权限请求
      type: object
      description: 分配权限输入
      properties:
        permission_ids:
          type: array
          description: 权限ID列表
          items:
            type: string
            minLength: 26
            maxLength: 26
          example: ["01HQZX123456789ABCDEFGHIJ", "01HQZX123456789ABCDEFGHIK"]
      required:
        - permission_ids

    # 角色信息
...
    UserRoleInfo:
      title: 用户角色信息
      type: object
      description: 用户角色信息
      properties:
        id:
          type: string
          description: 用户角色关联ID
          example: "01HQZX123456789ABCDEFGHIJ"
        role:
          $ref: '#/components/schemas/RoleInfo'
        granted_by:
          type: string
          description: 授权者ID
          example: "01HQZX123456789ABCDEFGHIJ"
        granted_at:
          type: string
          format: date-time
          description: 授权时间
          example: "2024-01-01T00:00:00Z"
        expires_at:
          type: string
          format: date-time
          description: 过期时间
          example: "2024-12-31T23:59:59Z"
        status:
          type: string
          description: 状态
          enum: [active, inactive, expired]
          example: "active"
        remark:
          type: string
          description: 备注
          example: "临时分配管理员权限"
      required:
        - id
        - role
        - granted_at
        - status

    # 权限管理相关Schema定义

    # 创建权限输入
...
    RoleResponse:
      title: 角色响应
      type: object
      description: 角色响应
      properties:
        code:
          type: integer
          description: API状态码（0=成功，非0=失败）
          example: 0
        message:
          type: string
          description: 响应消息
          example: "获取角色成功"
        data:
          $ref: '#/components/schemas/RoleInfo'
        errors:
          type: array
          description: 错误详情列表
          items:
            $ref: '#/components/schemas/ErrorDetail'
        timestamp:
          type: integer
          format: int64
          description: 时间戳
          example: 1640995200
        request_id:
          type: string
          description: 请求ID
          example: "req-123456"
      required:
        - code
        - message
        - data
        - timestamp

    # 角色列表响应
    RolesResponse:
      title: 角色列表响应
      type: object
      description: 角色列表响应
      properties:
        code:
          type: integer
          description: API状态码（0=成功，非0=失败）
          example: 0
        message:
          type: string
          description: 响应消息
          example: "获取角色列表成功"
        data:
          type: object
          description: 角色列表数据
          properties:
            roles:
              type: array
              description: 角色列表
              items:
                $ref: '#/components/schemas/RoleInfo'
            total:
              type: integer
              format: int64
              description: 总记录数
              example: 10
            page:
              type: integer
              description: 当前页码
              example: 1
            page_size:
              type: integer
              description: 每页数量
              example: 20
            total_pages:
              type: integer
              description: 总页数
              example: 1
          required:
            - roles
            - total
            - page
            - page_size
            - total_pages
        errors:
          type: array
          description: 错误详情列表
          items:
            $ref: '#/components/schemas/ErrorDetail'
        timestamp:
          type: integer
          format: int64
          description: 时间戳
          example: 1640995200
        request_id:
          type: string
          description: 请求ID
          example: "req-123456"
      required:
        - code
        - message
        - data
        - timestamp

    # 用户角色列表响应
    UserRolesResponse:
      title: 用户角色列表响应
      type: object
      description: 用户角色列表响应
      properties:
        code:
          type: integer
          description: API状态码（0=成功，非0=失败）
          example: 0
        message:
          type: string
          description: 响应消息
          example: "获取用户角色成功"
        data:
          type: array
          description: 用户角色列表
          items:
            $ref: '#/components/schemas/UserRoleInfo'
        errors:
          type: array
          description: 错误详情列表
          items:
            $ref: '#/components/schemas/ErrorDetail'
        timestamp:
          type: integer
          format: int64
          description: 时间戳
          example: 1640995200
        request_id:
          type: string
          description: 请求ID
          example: "req-123456"
      required:
        - code
        - message
        - data
        - timestamp

    # 角色权限响应
    RolePermissionsResponse:
      title: 角色权限响应
      type: object
      description: 角色权限响应
      properties:
        code:
          type: integer
          description: API状态码（0=成功，非0=失败）
          example: 0
        message:
          type: string
          description: 响应消息
          example: "获取角色权限成功"
        data:
          $ref: '#/components/schemas/RolePermissionInfo'
        errors:
          type: array
          description: 错误详情列表
          items:
            $ref: '#/components/schemas/ErrorDetail'
        timestamp:
          type: integer
          format: int64
          description: 时间戳
          example: 1640995200
        request_id:
          type: string
          description: 请求ID
          example: "req-123456"
      required:
        - code
        - message
        - data
        - timestamp

    # 权限响应
...
    AssignRoleMenusInput:
      title: 分配角色菜单请求
      type: object
      description: 为角色分配菜单的输入
      properties:
        menu_ids:
          type: array
          description: 菜单ID列表
          items:
            type: string
            minLength: 26
            maxLength: 26
          example: ["01HQZX123456789ABCDEFGHIJ", "01HQZX123456789ABCDEFGHIK"]
      required:
        - menu_ids

    # 角色菜单输出
...
    RoleMenuResponse:
      title: 角色菜单响应
      type: object
      description: 角色菜单响应
      properties:
        code:
          type: integer
          description: API状态码（0=成功，非0=失败）
          example: 0
        message:
          type: string
          description: 响应消息
          example: "获取角色菜单成功"
        data:
          $ref: '#/components/schemas/RoleMenuOutput'
        errors:
          type: array
          description: 错误详情列表
          items:
            $ref: '#/components/schemas/ErrorDetail'
        timestamp:
          type: integer
          format: int64
          description: 时间戳
          example: 1640995200
        request_id:
          type: string
          description: 请求ID
          example: "req-123456"
      required:
        - code
        - message
        - data
        - timestamp

    # 用户菜单响应
    UserMenuResponse:
      title: 用户菜单响应
      type: object
      description: 用户菜单响应
      properties:
        code:
          type: integer
          description: API状态码（0=成功，非0=失败）
          example: 0
        message:
          type: string
          description: 响应消息
          example: "获取用户菜单成功"
        data:
          $ref: '#/components/schemas/UserMenuOutput'
        errors:
          type: array
          description: 错误详情列表
          items:
            $ref: '#/components/schemas/ErrorDetail'
        timestamp:
          type: integer
          format: int64
          description: 时间戳
          example: 1640995200
        request_id:
          type: string
          description: 请求ID
          example: "req-123456"
      required:
        - code
        - message
        - data
        - timestamp

    # 角色菜单权限响应
    RoleMenuPermissionResponse:
      title: 角色菜单权限响应
      type: object
      description: 角色菜单权限响应
      properties:
        code:
          type: integer
          description: API状态码（0=成功，非0=失败）
          example: 0
        message:
          type: string
          description: 响应消息
          example: "获取角色菜单权限成功"
        data:
          $ref: '#/components/schemas/RoleMenuPermissionOutput'
        errors:
          type: array
          description: 错误详情列表
          items:
            $ref: '#/components/schemas/ErrorDetail'
        timestamp:
          type: integer
          format: int64
          description: 时间戳
          example: 1640995200
        request_id:
          type: string
          description: 请求ID
          example: "req-123456"
      required:
        - code
        - message
        - data
        - timestamp
  parameters:
    PageParam:
      name: page
      in: query
      description: 页码（从1开始）
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      name: page_size
      in: query
      description: 每页数量
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    UserIdParam:
      name: id
      in: path
      description: 用户ID
      required: true
      schema:
        type: string